pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval"
version = "0.1.0"
description = "LLM tool-calling evaluation toolkit with DeepEval"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "deepeval>=3.9.4",
    "litellm>=1.80.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "config-morpher @ git+https://github.com/kenyo3026/configmorpher.git",
    "rich>=14.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[project.scripts]
llm-eval = "llm_eval.cli:main"

# src-layout configuration
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
llm_eval = ["**/*.yaml", "**/*.json"]

[tool.setuptools]
include-package-data = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--verbose"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["deepeval.*", "litellm.*", "config_morpher.*"]
ignore_missing_imports = true
